implemented authorization

+++ src/api/notes/handler.js
@@ -16,8 +16,14 @@ class NotesHandler {
     try {
       this._validator.validateNotePayload(request.payload);
       const { title = "untitled", body, tags } = request.payload;
+      const { id: credentialId } = request.auth.credentials;
 
-      const noteId = await this._service.addNote({ title, body, tags });
+      const noteId = await this._service.addNote({
+        title,
+        body,
+        tags,
+        owner: credentialId,
+      });
 
       const response = h.response({
         status: "success",
@@ -50,8 +56,9 @@ class NotesHandler {
     }
   }
 
-  async getNotesHandler() {
-    const notes = await this._service.getNotes();
+  async getNotesHandler(request) {
+    const { id: credentialId } = request.auth.credentials;
+    const notes = await this._service.getNotes(credentialId);
     return {
       status: "success",
       data: {
@@ -63,6 +70,9 @@ class NotesHandler {
   async getNoteByIdHandler(request, h) {
     try {
       const { id } = request.params;
+      const { id: credentialId } = request.auth.credentials;
+
+      await this._service.verifyNoteOwner(id, credentialId);
       const note = await this._service.getNoteById(id);
       return {
         status: "success",
@@ -91,12 +101,14 @@ class NotesHandler {
     }
   }
 
-  putNoteByIdHandler(request, h) {
+  async putNoteByIdHandler(request, h) {
     try {
       this._validator.validateNotePayload(request.payload);
       const { id } = request.params;
+      const { id: credentialId } = request.auth.credentials;
 
-      this._service.editNoteById(id, request.payload);
+      await this._service.verifyNoteOwner(id, credentialId);
+      await this._service.editNoteById(id, request.payload);
 
       return {
         status: "success",
@@ -122,11 +134,13 @@ class NotesHandler {
     }
   }
 
-  deleteNoteByIdHandler(request, h) {
+  async deleteNoteByIdHandler(request, h) {
     try {
       const { id } = request.params;
+      const { id: credentialId } = request.auth.credentials;
 
-      this._service.deleteNoteById(id);
+      await this._service.verifyNoteOwner(id, credentialId);
+      await this._service.deleteNoteById(id);
 
       return {
         status: "success",
